feat: moving/organizing DTO folders
