+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/bussines/FileStorageService.java
@@ -3,11 +3,8 @@ package br.com.norteautopecas.painel_administrativo_backend.bussines;
 import java.nio.file.*;
 
 import br.com.norteautopecas.painel_administrativo_backend.config.FileStorageConfig;
-import br.com.norteautopecas.painel_administrativo_backend.infra.dto.ListarFilesPorTicketIdDTO;
-import br.com.norteautopecas.painel_administrativo_backend.infra.dto.TicketDivergenciaDetailsDTO;
-import br.com.norteautopecas.painel_administrativo_backend.infra.dto.UploadFileResponseDTO;
-import br.com.norteautopecas.painel_administrativo_backend.infra.entity.StoreInformation;
-import br.com.norteautopecas.painel_administrativo_backend.infra.entity.TicketDivergencia;
+import br.com.norteautopecas.painel_administrativo_backend.infra.dto.files.ListarFilesPorTicketIdDTO;
+import br.com.norteautopecas.painel_administrativo_backend.infra.dto.files.UploadFileResponseDTO;
 import br.com.norteautopecas.painel_administrativo_backend.infra.entity.TicketFiles;
 import br.com.norteautopecas.painel_administrativo_backend.infra.exception.FileNotFoundException;
 import br.com.norteautopecas.painel_administrativo_backend.infra.exception.FileStorageException;
