+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/infra/dto/divergencia/TicketDivergenciaCreateDTO.java
@@ -1,16 +1,12 @@
-package br.com.norteautopecas.painel_administrativo_backend.infra.dto;
+package br.com.norteautopecas.painel_administrativo_backend.infra.dto.divergencia;
 
-
+import br.com.norteautopecas.painel_administrativo_backend.infra.dto.produto.ProdutoCreateDTO;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.*;
 
 import java.util.List;
 
-public record TicketGarantiaCreateDTO(
-        @JsonProperty("nome_cliente")
-        @NotBlank(message = "o Campo nome do cliente é obrigatório")
-        String nomeCliente,
-
+public record TicketDivergenciaCreateDTO(
         @NotNull(message = "o Campo loja é obrigatório")
         @Min(value = 100, message = "A loja deve ter no mínimo 3 dígitos")
         @Max(value = 999, message = "A loja deve ter no máximo 3 dígitos")
@@ -39,5 +35,4 @@ public record TicketGarantiaCreateDTO(
         @NotEmpty(message = "Não é possível criar um ticket sem produtos")
         List<ProdutoCreateDTO> produtos
 ) {
-
 }
