+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/infra/dto/lojas/ListStoreByUserDTO.java
@@ -1,0 +1,8 @@
+package br.com.norteautopecas.painel_administrativo_backend.infra.dto.lojas;
+
+public record ListStoreByUserDTO(
+        Long idUser,
+        String nome,
+        Integer loja
+) {
+}
