+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/infra/dto/lojas/StoreRegistrationDetailsDTO.java
@@ -1,4 +1,4 @@
-package br.com.norteautopecas.painel_administrativo_backend.infra.dto;
+package br.com.norteautopecas.painel_administrativo_backend.infra.dto.lojas;
 
 import br.com.norteautopecas.painel_administrativo_backend.infra.entity.User;
 import com.fasterxml.jackson.annotation.JsonAlias;
