+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/infra/dto/message/TicketMessageCreateDTO.java
@@ -1,6 +1,5 @@
-package br.com.norteautopecas.painel_administrativo_backend.infra.dto;
+package br.com.norteautopecas.painel_administrativo_backend.infra.dto.message;
 
-import br.com.norteautopecas.painel_administrativo_backend.infra.dto.users.UserRegistrationDataDTO;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
