+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/infra/dto/message/TicketMessageDetailsDTO.java
@@ -1,4 +1,4 @@
-package br.com.norteautopecas.painel_administrativo_backend.infra.dto;
+package br.com.norteautopecas.painel_administrativo_backend.infra.dto.message;
 
 
 import br.com.norteautopecas.painel_administrativo_backend.infra.dto.users.UserRegistrationDataDTO;
