+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/infra/dto/status_historico/TicketStatusHistoricoCreateDTO.java
@@ -1,4 +1,4 @@
-package br.com.norteautopecas.painel_administrativo_backend.infra.dto;
+package br.com.norteautopecas.painel_administrativo_backend.infra.dto.status_historico;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import jakarta.validation.constraints.NotBlank;
