+++ src/main/java/br/com/norteautopecas/painel_administrativo_backend/infra/dto/status_historico/TicketStatusHistoricoListByIdDTO.java
@@ -1,7 +1,6 @@
-package br.com.norteautopecas.painel_administrativo_backend.infra.dto;
+package br.com.norteautopecas.painel_administrativo_backend.infra.dto.status_historico;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record TicketStatusHistoricoListByIdDTO(
